security v

+++ src/main/java/by/iba/kazakov/device_verification/services/serviceInterfaces/UserService.java
@@ -3,5 +3,5 @@ package by.iba.kazakov.device_verification.services.serviceInterfaces;
 import by.iba.kazakov.device_verification.models.User;
 import by.iba.kazakov.device_verification.services.serviceInterfaces.CrudService;
 
-public interface UserService extends CrudService<User, Integer> {
+public interface UserService extends CrudService<User, Long> {
 }
